Fix extension support issue crashing the browser, and update the appservice name
